feat: added base table
